+++ theories/sets.v
@@ -1169,6 +1169,46 @@ Proof.
   intros xs. exists (fresh xs). split; [set_solver|]. apply infinite_is_fresh.
 Qed.
 
+Lemma dec_pred_finite {A} (P : A → Prop) {Hdec : ∀ x, Decision (P x)} :
+  pred_finite P ↔ ∃ (xs : list A), ∀ x, P x ↔ x ∈ xs.
+Proof.
+  split; intros [xs Hxs]; [|exists xs; set_solver].
+  cut (∀ n, ∃ ys, (∀ x, P x → x ∈ ys ++ drop n xs) ∧ (∀ x, x ∈ ys → P x)).
+  { intros H. specialize (H (length xs)) as (ys & H1 & H2).
+    rewrite drop_all, app_nil_r in H1. exists ys. set_solver. }
+  intros n. induction n as [ | n (ys & IH1 & IH2)].
+  { exists []. rewrite drop_0. set_solver. }
+  destruct (decide (n < length xs)) as [[y Hn]%lookup_lt_is_Some | ?].
+  { destruct (decide (P y)) as [Hy | Hy].
+    { exists (ys ++ [y]). pose proof (assoc app) as <-. cbn.
+      rewrite <-drop_S; set_solver. }
+    { exists ys. split; [|done]. intros x Hx.
+      specialize (IH1 x Hx) as [? | Hx_elem_of]%elem_of_app; [set_solver|].
+      erewrite drop_S in Hx_elem_of; set_solver. } }
+  { exists ys. revert IH1. rewrite !drop_ge, app_nil_r; [done|lia..]. }
+Qed.
+
+Lemma finite_sig_pred_finite {A} (P : A → Prop) `{Hfin : Finite (sig P)} :
+  pred_finite P.
+Proof.
+  exists (proj1_sig <$> enum _). intros x px.
+  apply elem_of_list_fmap_1_alt with (x ↾ px); [apply elem_of_enum|]; done.
+Qed.
+
+Lemma pred_finite_arg2 {A B} (P : A → B → Prop) :
+  pred_finite (uncurry P) → ∀ x, pred_finite (P x).
+Proof.
+  intros [xys ?] x. exists (xys.*2). intros y ?.
+  apply elem_of_list_fmap_1_alt with (x, y); by auto.
+Qed.
+
+Lemma pred_finite_arg1 {A B} (P : A → B → Prop) :
+  pred_finite (uncurry P) → ∀ y, pred_finite (flip P y).
+Proof.
+  intros [xys ?] y. exists (xys.*1). intros x ?.
+  apply elem_of_list_fmap_1_alt with (x, y); by auto.
+Qed.
+
 (** Sets of sequences of natural numbers *)
 (* The set [seq_seq start len] of natural numbers contains the sequence
 [start, start + 1, ..., start + (len-1)]. *)